Update updateTaskContainer function and add editTask and deleteTask functions
Updated the updateTaskContainer function to clear the tasksContainer inner HTML before populating it with task data. Added onclick attributes to the Edit and Delete buttons to call the editTask and deleteTask functions respectively. The deleteTask function removes the task element from the DOM and also removes the corresponding task data from the taskData array. The editTask function updates the currentTask object with the data of the task to be edited.